Fix reference to potentially missing gems from gemspec
mollie.gemspec `import`ed Mollie/API/Client, which then did `require
“rest_client”`. This is effectively a circular reference as it’s the
gemspec file that indicates this gem is required, and in a clean gemset
where rest-client is not installed the gem cannot be built.

This fix moves the version constant into a separate file with no
requirements of its own and references that from the gemspec,
preserving the existing API but removing the circular reference.

I’ve also added a Gemfile to make it easier to install the required
dependencies for working on the gem.